+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Local setup to test package
 - Add webpage documentation
 - Date enter in input of datepicker
+- Define datepicker date format
 
 ### Fixed
 - Duplicate main component